pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "finarg"
version = "0.1.0"
description = "AI-powered financial agent for Argentina/LATAM"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
authors = [{ name = "Gantalf" }]
keywords = ["ai", "finance", "crypto", "argentina", "agent", "tui"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Office/Business :: Financial",
]
dependencies = [
    "textual>=0.47",
    "rich>=13.0",
    "httpx>=0.27",
    "anthropic>=0.40",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "python-dotenv>=1.0",
    "tenacity>=8.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "ruff>=0.4",
]

[project.scripts]
finarg = "finarg.cli:main"

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.pytest.ini_options]
asyncio_mode = "auto"
